Reduce requested simulation variable size (flaot vs double)

+++ src/SkyConnect/src/SimConnectPrimaryFlightControl.h
@@ -39,9 +39,9 @@
 #pragma pack(push, 1)
 struct SimConnectPrimaryFlightControl
 {
-    double rudderPosition;
-    double elevatorPosition;
-    double aileronPosition;
+    float rudderPosition;
+    float elevatorPosition;
+    float aileronPosition;
 
     inline PrimaryFlightControlData toPrimaryFlightControlData() const noexcept
     {
